+++ ChangeLog.md
@@ -8,9 +8,12 @@
 
 ## 10/20/2022 3.2.1
 
+- ModelValidator - Fix bug during reporting real exampleJsonPath
+- ModelValidator - Add LRO header validation for data-planes swagger
 - GenerateExamples - Add title and operationId to example
 - GenerateExamples - Generate subscriptionId in guid format
 - ModelValidator - Add type validation for additionalProperties whose type is specified as an object
+- Bugfix: FileLoader - isUnderFileRoot function supports validation for absolute path
 - API Scenario
   - Add roleAssignment step
   - Bugfix: Use content-type specified in swagger